fix(admin): Handle TotalStorageLimit in MB instead of Bytes in UI and fix InvalidCastException on UsageMetrics

+++ WebApp/Backend/PlanlamaApp.Api/Controllers/AdminController.cs
@@ -122,13 +122,13 @@ namespace PlanlamaApp.Api.Controllers
         {
             var stats = new
             {
-                TotalUsers = await dbConnection.ExecuteScalarAsync<int>("SELECT COUNT(*) FROM Users WHERE IsActive = true;"),
-                TotalUsersDeleted = await dbConnection.ExecuteScalarAsync<int>("SELECT COUNT(*) FROM Users WHERE IsActive = false;"),
-                PremiumUsers = await dbConnection.ExecuteScalarAsync<int>("SELECT COUNT(*) FROM Users WHERE SubscriptionPlan = 'premium' AND IsActive = true;"),
-                PremiumUsersDeleted = await dbConnection.ExecuteScalarAsync<int>("SELECT COUNT(*) FROM Users WHERE SubscriptionPlan = 'premium' AND IsActive = false;"),
-                TotalWorkspaces = await dbConnection.ExecuteScalarAsync<int>("SELECT COUNT(*) FROM Workspaces WHERE IsActive = true;"),
-                TotalWorkspacesDeleted = await dbConnection.ExecuteScalarAsync<int>("SELECT COUNT(*) FROM Workspaces WHERE IsActive = false;"),
-                TotalTasks = await dbConnection.ExecuteScalarAsync<int>("SELECT COUNT(*) FROM TaskItems;")
+                TotalUsers = await dbConnection.ExecuteScalarAsync<long>("SELECT COUNT(*) FROM Users WHERE IsActive = true;"),
+                TotalUsersDeleted = await dbConnection.ExecuteScalarAsync<long>("SELECT COUNT(*) FROM Users WHERE IsActive = false;"),
+                PremiumUsers = await dbConnection.ExecuteScalarAsync<long>("SELECT COUNT(*) FROM Users WHERE SubscriptionPlan = 'premium' AND IsActive = true;"),
+                PremiumUsersDeleted = await dbConnection.ExecuteScalarAsync<long>("SELECT COUNT(*) FROM Users WHERE SubscriptionPlan = 'premium' AND IsActive = false;"),
+                TotalWorkspaces = await dbConnection.ExecuteScalarAsync<long>("SELECT COUNT(*) FROM Workspaces WHERE IsActive = true;"),
+                TotalWorkspacesDeleted = await dbConnection.ExecuteScalarAsync<long>("SELECT COUNT(*) FROM Workspaces WHERE IsActive = false;"),
+                TotalTasks = await dbConnection.ExecuteScalarAsync<long>("SELECT COUNT(*) FROM TaskItems;")
             };
             return Ok(stats);
         }
